combustion/gnns/requirements.txt
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
# 
#     http://www.apache.org/licenses/LICENSE-2.0
# 
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

h5py=='3.4.0'
matplotlib=='3.4.3'
networkx=='2.6.3'
numpy=='1.21.2'
plotly=='5.3.1'
pytorch-lightning=='1.5.7'
PyYAML=='0.5.4'
randomname=='0.1.5'
torch=='1.9.0+cu111'
torchmetrics=='0.5.1'
torch_optimizer=='0.1.0'